backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.8.0rc2
Django==4.2.10
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==6.2.5
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.2.1
